add support for new WebAssembly.Instance(moduleObject[, importObject])
