+++ lib/Runtime/Library/WasmLibrary.h
@@ -35,27 +35,6 @@ namespace Js
         static Var WasmDeferredParseExternalThunk(RecyclableObject* function, CallInfo callInfo, ...);
 
     private:
-        static void WasmFunctionGenerateBytecode(AsmJsScriptFunction* func, bool propagateError);
-        static void WasmLoadDataSegs(Wasm::WasmModule * wasmModule, Var* heap, ScriptContext* ctx);
-        static void WasmLoadFunctions(Wasm::WasmModule * wasmModule, ScriptContext* ctx, Var* moduleMemoryPtr, Var* exportObj, Var* localModuleFunctions, bool* hasAnyLazyTraps);
-        static void WasmBuildObject(Wasm::WasmModule * wasmModule, ScriptContext* ctx, Var exportsNamespace, Var* heap, Var* exportObj, bool* hasAnyLazyTraps, Var* localModuleFunctions, Var* importFunctions);
-        static void WasmLoadImports(Wasm::WasmModule * wasmModule, ScriptContext* ctx, Var* importFunctions, Var moduleEnv, Var ffi);
-        static void WasmLoadGlobals(Wasm::WasmModule * wasmModule, ScriptContext* ctx, Var moduleEnv, Var ffi);
-        static void WasmLoadIndirectFunctionTables(Wasm::WasmModule * wasmModule, ScriptContext* ctx, Var** indirectFunctionTables, Var* localModuleFunctions, Var* importFunctions);
-        static Var GetFunctionObjFromFunctionIndex(Wasm::WasmModule * wasmModule, ScriptContext* ctx, uint32 funcIndex, Var* localModuleFunctions, Var* importFunctions);
-
-        static Var LoadWasmScript(
-            ScriptContext* scriptContext,
-            const char16* script,
-            Js::Var bufferSrc,
-            SRCINFO const * pSrcInfo,
-            CompileScriptException * pse,
-            Utf8SourceInfo** ppSourceInfo,
-            const uint lengthBytes,
-            Js::Var ffi,
-            Js::Var* start = nullptr
-        );
-        static char16* lastWasmExceptionMessage;
 #endif
     };
 }
