+++ lib/WasmReader/Chakra.WasmReader.vcxproj
@@ -36,7 +36,6 @@
     </ClCompile>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="$(MSBuildThisFileDirectory)WasmModule.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)WasmBytecodeGenerator.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)WasmDataSegment.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)WasmFunctionInfo.cpp" />
@@ -50,7 +49,6 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="WasmGlobal.h" />
-    <ClInclude Include="WasmModule.h" />
     <ClInclude Include="WasmBytecodeGenerator.h" />
     <ClInclude Include="WasmDataSegment.h" />
     <ClInclude Include="WasmFunctionInfo.h" />
